apps/webserver/app/requirements.txt
flask==2.0.2
python-dotenv==0.19.2
mcstatus==6.5.0
